Use unified regression test binary
