+++ tests/CMakeLists.txt
@@ -39,9 +39,6 @@ FOREACH(xml ${PRIMITIVE_TESTS})
   ENDFOREACH()
 ENDFOREACH()
 
-ADD_SUBDIRECTORY(regression)
-
-
 IF (EMBREE_TESTING_INSTALL_TESTS)
   # test resources
   INSTALL(DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/primitives" DESTINATION "${CMAKE_INSTALL_TESTDIR}/tests" COMPONENT testing PATTERN "*.py" EXCLUDE)
